refactor(swift): move the mock URLProtocol into OSNTesting
A `URLProtocol` subclass that answers requests from a settable handler was
duplicated, near enough line for line, across five test files:
`LoginMockURLProtocol` (OSNAuthTests, used by two suites), `MockURLProtocol`
(OSNKitTests), `MiddlewareMockURLProtocol` (PulseAPITests) and
`MusubiMockURLProtocol` (MusubiFeatureTests). Each copy was deliberate — SwiftPM
shares nothing between test targets — but `OSNTesting` is a *source* target all
five already depend on, which is the seam that was missing.

One copy now lives in `Sources/OSNTesting/MockURLProtocol.swift`, along with the
`makeMockSession()` helper the five files each spelled identically. The cookie
capture is carried over verbatim: a custom `URLProtocol` does not run the real
transport's `Set-Cookie` extraction, and both `TokenRefresher` and
`PasskeyLoginClient` verify the rotated session cookie actually landed in the
jar, so without it they fail against a mock that otherwise looks correct.

Five statics become one, and Swift Testing runs suites and targets
concurrently — so the shared handler is only safe because every suite touching
it already carries `.keychainSerializing`, which serialises across suites and
targets. That was checked rather than assumed: OSNSessionTests,
PasskeyLoginClientTests, BearerTokenMiddlewareTests and MusubiFeatureTests'
FetchPasskeysTests all declare it, and TokenRefresherTests extends
OSNKitTests' KeychainSerialTests, which does too. The requirement is written
into the mock's own doc comment so the next suite that uses it knows.

No production behaviour changes. `scripts/changeset-required.sh` reports `skip`
— every changed path is on the allowlist.

Closes #794. Also closes the finding it was filed from, xchromo/osn-tracker#451.
