+++ shared/swift/OSNShared/Sources/OSNTesting/OSNTesting.swift
@@ -1,3 +1,8 @@
-// OSNTesting: fixtures/fakes for the app test targets. Empty in A0 — depends
-// on OSNKit since fakes here stand in for its API client/session types.
+// OSNTesting: fixtures, fakes and traits shared by the app test targets.
+//
+// A source target rather than a test target, because SwiftPM shares source
+// targets across test targets and shares nothing between test targets — which
+// is why `MockURLProtocol` lived in five copies before it moved here.
+//
+// Depends on OSNKit since fakes here stand in for its client/session types.
 import OSNKit
